feat:【PAC模板】流水线模板支持PAC特性 #11414 基本结构、类、方法设计

+++ src/backend/ci/core/common/common-pipeline/src/main/kotlin/com/tencent/devops/common/pipeline/template/PipelineTemplateInfo.kt
@@ -25,11 +25,11 @@ data class PipelineTemplateInfo(
     @get:Schema(title = "是否开启PAC", required = true)
     val enablePac: Boolean,
     @get:Schema(title = "最新版本号", required = true)
-    val lastedVersion: Long,
+    val lastedVersion: Int,
     @get:Schema(title = "最新版本名称", required = true)
     val lastedVersionName: String,
     @get:Schema(title = "最新设置版本号", required = true)
-    val lastedSettingVersion: Long,
+    val lastedSettingVersion: Int,
     @get:Schema(title = "模板来源", required = true)
     val source: PipelineTemplateSource,
     @get:Schema(title = "是否关联研发商店", required = true)
